requirements.txt
numpy~=1.24.2
matplotlib~=3.7.1
python-dotenv~=1.0.0